refactor backend to be more modular
- separate app, model, main
  - resolves #19
- add pytest tests of model and app
  - resolves #20
