+++ backend/src/circuit_seq_server/__init__.py
@@ -1,1 +1,3 @@
 __version__ = "0.0.1"
+
+from circuit_seq_server.app import create_app
